Working version without save
Animations ok,
Move the InputText to the side or up at the top and down at the bottom
to come back to main view
Top to send thougths, bottom to read past thougths.
But for now nothing is record so nothing to read.
To fix: smiley press enable in animation. Have to wait the end of the
animation to enable it.
To add: data management to save/delete/read thoughts.